Docs: Examples for RFC6587 (LF and NUL trailers)
